refactor: tracer pkg
